Introduce MimeType archive matcher and aggregate matcher
In order to recognise the type of archive when it's an uploaded file, and does not have a file name extension, the mime type matcher can be used to guess the mime type and match to the appropriate adapter.

The aggregate matcher allows chaining of matchers so that multiple strategies can be attempted.

Signed-off-by: George Steel <george@net-glue.co.uk>